Rename string.h -> String.h
